refactor(webapp): regroup allocated-transactions under shared dir
Merge allocated-transactions-dialog/ + create-allocated-transaction-dialog/
into allocated-transactions/{details-dialog,create-dialog}/. Strip redundant
entity prefix from filenames (dialog.ts, bottom-sheet.ts) — class names and
exports unchanged.
