+++ frontend/projects/webapp/src/app/feature/budget/budget-details/allocated-transactions/create-dialog/budget-period-date-constraints.ts
