+++ frontend/projects/webapp/src/app/feature/budget/budget-details/allocated-transactions/create-dialog/dialog.spec.ts
@@ -3,27 +3,23 @@ import { TestBed } from '@angular/core/testing';
 import { provideZonelessChangeDetection, signal } from '@angular/core';
 import { provideAnimationsAsync } from '@angular/platform-browser/animations/async';
 import { provideNativeDateAdapter } from '@angular/material/core';
-import {
-  MAT_BOTTOM_SHEET_DATA,
-  MatBottomSheetRef,
-} from '@angular/material/bottom-sheet';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { provideTranslocoForTest } from '@app/testing/transloco-testing';
 import { CurrencyConverterService } from '@core/currency';
 import { FeatureFlagsService } from '@core/feature-flags';
 import { UserSettingsStore } from '@core/user-settings';
 import type { SupportedCurrency } from 'pulpe-shared';
-import { CreateAllocatedTransactionBottomSheet } from './create-allocated-transaction-bottom-sheet';
-import type { CreateAllocatedTransactionDialogData } from './create-allocated-transaction-dialog';
-
-const TEST_BUDGET_LINE_ID = '11111111-1111-4111-8111-111111111111';
-const TEST_BUDGET_ID = '22222222-2222-4222-8222-222222222222';
+import {
+  CreateAllocatedTransactionDialog,
+  type CreateAllocatedTransactionDialogData,
+} from './dialog';
 
 const createDialogData = (
   overrides: Partial<CreateAllocatedTransactionDialogData['budgetLine']> = {},
 ): CreateAllocatedTransactionDialogData => ({
   budgetLine: {
-    id: TEST_BUDGET_LINE_ID,
-    budgetId: TEST_BUDGET_ID,
+    id: '00000000-0000-4000-8000-0000000000b1',
+    budgetId: '00000000-0000-4000-8000-000000000456',
     name: 'Assurance maladie',
     amount: 385,
     kind: 'expense',
@@ -38,24 +34,24 @@ const createDialogData = (
   payDayOfMonth: null,
 });
 
-describe('CreateAllocatedTransactionBottomSheet', () => {
-  let component: CreateAllocatedTransactionBottomSheet;
-  let mockBottomSheetRef: { dismiss: ReturnType<typeof vi.fn> };
+describe('CreateAllocatedTransactionDialog', () => {
+  let component: CreateAllocatedTransactionDialog;
+  let mockDialogRef: { close: ReturnType<typeof vi.fn> };
   let dialogData: CreateAllocatedTransactionDialogData;
 
   beforeEach(async () => {
-    mockBottomSheetRef = { dismiss: vi.fn() };
+    mockDialogRef = { close: vi.fn() };
     dialogData = createDialogData();
 
     await TestBed.configureTestingModule({
-      imports: [CreateAllocatedTransactionBottomSheet],
+      imports: [CreateAllocatedTransactionDialog],
       providers: [
         provideZonelessChangeDetection(),
         provideAnimationsAsync(),
         provideNativeDateAdapter(),
         ...provideTranslocoForTest(),
-        { provide: MAT_BOTTOM_SHEET_DATA, useValue: dialogData },
-        { provide: MatBottomSheetRef, useValue: mockBottomSheetRef },
+        { provide: MAT_DIALOG_DATA, useValue: dialogData },
+        { provide: MatDialogRef, useValue: mockDialogRef },
         {
           provide: CurrencyConverterService,
           useValue: {
@@ -71,12 +67,12 @@ describe('CreateAllocatedTransactionBottomSheet', () => {
     }).compileComponents();
 
     component = TestBed.createComponent(
-      CreateAllocatedTransactionBottomSheet,
+      CreateAllocatedTransactionDialog,
     ).componentInstance;
   });
 
   describe('submit', () => {
-    it('should dismiss with transaction data when form is valid', async () => {
+    it('should close with transaction data when form is valid', async () => {
       const midMonth = new Date(
         new Date().getFullYear(),
         new Date().getMonth(),
@@ -91,10 +87,10 @@ describe('CreateAllocatedTransactionBottomSheet', () => {
 
       await component['submit']();
 
-      expect(mockBottomSheetRef.dismiss).toHaveBeenCalledWith(
+      expect(mockDialogRef.close).toHaveBeenCalledWith(
         expect.objectContaining({
-          budgetId: TEST_BUDGET_ID,
-          budgetLineId: TEST_BUDGET_LINE_ID,
+          budgetId: '00000000-0000-4000-8000-000000000456',
+          budgetLineId: '00000000-0000-4000-8000-0000000000b1',
           name: 'Consultation médecin',
           amount: 45.5,
           kind: 'expense',
@@ -103,7 +99,7 @@ describe('CreateAllocatedTransactionBottomSheet', () => {
       );
     });
 
-    it('should not dismiss when form is invalid', async () => {
+    it('should not close when form is invalid', async () => {
       component['model'].update((m) => ({
         ...m,
         name: '',
@@ -112,7 +108,7 @@ describe('CreateAllocatedTransactionBottomSheet', () => {
 
       await component['submit']();
 
-      expect(mockBottomSheetRef.dismiss).not.toHaveBeenCalled();
+      expect(mockDialogRef.close).not.toHaveBeenCalled();
     });
 
     it('should trim whitespace from name', async () => {
@@ -130,7 +126,7 @@ describe('CreateAllocatedTransactionBottomSheet', () => {
 
       await component['submit']();
 
-      expect(mockBottomSheetRef.dismiss).toHaveBeenCalledWith(
+      expect(mockDialogRef.close).toHaveBeenCalledWith(
         expect.objectContaining({ name: 'Courses' }),
       );
     });
@@ -150,7 +146,7 @@ describe('CreateAllocatedTransactionBottomSheet', () => {
 
       await component['submit']();
 
-      expect(mockBottomSheetRef.dismiss).toHaveBeenCalledWith(
+      expect(mockDialogRef.close).toHaveBeenCalledWith(
         expect.objectContaining({ amount: 42.5 }),
       );
     });
@@ -172,7 +168,7 @@ describe('CreateAllocatedTransactionBottomSheet', () => {
 
       await component['submit']();
 
-      expect(mockBottomSheetRef.dismiss).toHaveBeenCalledWith(
+      expect(mockDialogRef.close).toHaveBeenCalledWith(
         expect.objectContaining({ checkedAt: null }),
       );
     });
@@ -193,7 +189,7 @@ describe('CreateAllocatedTransactionBottomSheet', () => {
 
       await component['submit']();
 
-      const callArg = mockBottomSheetRef.dismiss.mock.calls[0][0];
+      const callArg = mockDialogRef.close.mock.calls[0][0];
       expect(callArg.checkedAt).toBeDefined();
       expect(typeof callArg.checkedAt).toBe('string');
       expect(() => new Date(callArg.checkedAt!)).not.toThrow();
@@ -215,17 +211,17 @@ describe('CreateAllocatedTransactionBottomSheet', () => {
 
       await component['submit']();
 
-      expect(mockBottomSheetRef.dismiss).toHaveBeenCalledWith(
+      expect(mockDialogRef.close).toHaveBeenCalledWith(
         expect.objectContaining({ checkedAt: null }),
       );
     });
   });
 
-  describe('close', () => {
-    it('should dismiss without data', () => {
-      component['close']();
+  describe('cancel', () => {
+    it('should close without data', () => {
+      component['cancel']();
 
-      expect(mockBottomSheetRef.dismiss).toHaveBeenCalledWith();
+      expect(mockDialogRef.close).toHaveBeenCalledWith();
     });
   });
 
@@ -308,88 +304,6 @@ describe('CreateAllocatedTransactionBottomSheet', () => {
       ).toBe(true);
     });
   });
-
-  describe('date constraints', () => {
-    it('should set minDate and maxDate for current month budget', () => {
-      expect(component['minDate']).toBeDefined();
-      expect(component['maxDate']).toBeDefined();
-      expect(component['minDate']!.getTime()).toBeLessThanOrEqual(
-        component['maxDate']!.getTime(),
-      );
-    });
-
-    it('should set minDate and maxDate for past month budget', async () => {
-      const pastData: CreateAllocatedTransactionDialogData = {
-        ...createDialogData(),
-        budgetMonth: 1,
-        budgetYear: 2020,
-      };
-
-      const pastRef = { dismiss: vi.fn() };
-
-      await TestBed.resetTestingModule()
-        .configureTestingModule({
-          imports: [CreateAllocatedTransactionBottomSheet],
-          providers: [
-            provideZonelessChangeDetection(),
-            provideAnimationsAsync(),
-            provideNativeDateAdapter(),
-            ...provideTranslocoForTest(),
-            { provide: MAT_BOTTOM_SHEET_DATA, useValue: pastData },
-            { provide: MatBottomSheetRef, useValue: pastRef },
-          ],
-        })
-        .compileComponents();
-
-      const pastComponent = TestBed.createComponent(
-        CreateAllocatedTransactionBottomSheet,
-      ).componentInstance;
-
-      expect(pastComponent['minDate']).toBeDefined();
-      expect(pastComponent['maxDate']).toBeDefined();
-      expect(pastComponent['minDate']!.getMonth()).toBe(0);
-      expect(pastComponent['minDate']!.getFullYear()).toBe(2020);
-    });
-
-    it('should respect custom payDayOfMonth', async () => {
-      vi.useFakeTimers();
-      vi.setSystemTime(new Date(2025, 5, 27));
-
-      const customPayDayData: CreateAllocatedTransactionDialogData = {
-        ...createDialogData(),
-        budgetMonth: 7,
-        budgetYear: 2025,
-        payDayOfMonth: 25,
-      };
-
-      const customRef = { dismiss: vi.fn() };
-
-      await TestBed.resetTestingModule()
-        .configureTestingModule({
-          imports: [CreateAllocatedTransactionBottomSheet],
-          providers: [
-            provideZonelessChangeDetection(),
-            provideAnimationsAsync(),
-            provideNativeDateAdapter(),
-            ...provideTranslocoForTest(),
-            { provide: MAT_BOTTOM_SHEET_DATA, useValue: customPayDayData },
-            { provide: MatBottomSheetRef, useValue: customRef },
-          ],
-        })
-        .compileComponents();
-
-      const customComponent = TestBed.createComponent(
-        CreateAllocatedTransactionBottomSheet,
-      ).componentInstance;
-
-      expect(customComponent['minDate']).toBeDefined();
-      expect(customComponent['maxDate']).toBeDefined();
-      expect(customComponent['minDate']!.getDate()).toBe(25);
-      expect(customComponent['maxDate']!.getDate()).toBe(24);
-
-      vi.useRealTimers();
-    });
-  });
 });
 
 interface FlagsMock {
@@ -402,11 +316,11 @@ interface SettingsMock {
 interface ConverterMock {
   convertWithMetadata: ReturnType<typeof vi.fn>;
 }
-interface BottomSheetRefMock {
-  dismiss: ReturnType<typeof vi.fn>;
+interface DialogRefMock {
+  close: ReturnType<typeof vi.fn>;
 }
 
-function configureBottomSheetWithCurrency({
+function configureDialogWithCurrency({
   userCurrency,
   flagEnabled,
   showCurrencyPref = true,
@@ -415,7 +329,7 @@ function configureBottomSheetWithCurrency({
   flagEnabled: boolean;
   showCurrencyPref?: boolean;
 }) {
-  const bottomSheetRef: BottomSheetRefMock = { dismiss: vi.fn() };
+  const dialogRef: DialogRefMock = { close: vi.fn() };
   const flags: FlagsMock = {
     isMultiCurrencyEnabled: signal(flagEnabled),
   };
@@ -431,32 +345,30 @@ function configureBottomSheetWithCurrency({
   };
 
   TestBed.configureTestingModule({
-    imports: [CreateAllocatedTransactionBottomSheet],
+    imports: [CreateAllocatedTransactionDialog],
     providers: [
       provideZonelessChangeDetection(),
       provideAnimationsAsync(),
       provideNativeDateAdapter(),
       ...provideTranslocoForTest(),
-      { provide: MAT_BOTTOM_SHEET_DATA, useValue: createDialogData() },
-      { provide: MatBottomSheetRef, useValue: bottomSheetRef },
+      { provide: MAT_DIALOG_DATA, useValue: createDialogData() },
+      { provide: MatDialogRef, useValue: dialogRef },
       { provide: FeatureFlagsService, useValue: flags },
       { provide: UserSettingsStore, useValue: settings },
       { provide: CurrencyConverterService, useValue: converter },
     ],
   });
 
-  const fixture = TestBed.createComponent(
-    CreateAllocatedTransactionBottomSheet,
-  );
+  const fixture = TestBed.createComponent(CreateAllocatedTransactionDialog);
   const component = fixture.componentInstance;
-  return { fixture, component, bottomSheetRef, converter, settings, flags };
+  return { fixture, component, dialogRef, converter, settings, flags };
 }
 
-describe('CreateAllocatedTransactionBottomSheet — currency create rules', () => {
+describe('CreateAllocatedTransactionDialog — currency create rules', () => {
   beforeEach(() => TestBed.resetTestingModule());
 
   it('should initialize money slice with user currency', () => {
-    const { component } = configureBottomSheetWithCurrency({
+    const { component } = configureDialogWithCurrency({
       userCurrency: 'EUR',
       flagEnabled: true,
     });
@@ -465,11 +377,10 @@ describe('CreateAllocatedTransactionBottomSheet — currency create rules', () =
   });
 
   it('should call convertWithMetadata and include metadata in payload when currencies differ', async () => {
-    const { component, bottomSheetRef, converter } =
-      configureBottomSheetWithCurrency({
-        userCurrency: 'CHF',
-        flagEnabled: true,
-      });
+    const { component, dialogRef, converter } = configureDialogWithCurrency({
+      userCurrency: 'CHF',
+      flagEnabled: true,
+    });
 
     converter.convertWithMetadata.mockResolvedValue({
       convertedAmount: 108.97,
@@ -500,8 +411,8 @@ describe('CreateAllocatedTransactionBottomSheet — currency create rules', () =
       'EUR',
       'CHF',
     );
-    expect(bottomSheetRef.dismiss).toHaveBeenCalledTimes(1);
-    const dto = bottomSheetRef.dismiss.mock.calls[0][0];
+    expect(dialogRef.close).toHaveBeenCalledTimes(1);
+    const dto = dialogRef.close.mock.calls[0][0];
     expect(dto.amount).toBe(108.97);
     expect(dto.originalAmount).toBe(100);
     expect(dto.originalCurrency).toBe('EUR');
@@ -510,11 +421,10 @@ describe('CreateAllocatedTransactionBottomSheet — currency create rules', () =
   });
 
   it('should block submit and set conversionError when convertWithMetadata throws', async () => {
-    const { component, bottomSheetRef, converter } =
-      configureBottomSheetWithCurrency({
-        userCurrency: 'CHF',
-        flagEnabled: true,
-      });
+    const { component, dialogRef, converter } = configureDialogWithCurrency({
+      userCurrency: 'CHF',
+      flagEnabled: true,
+    });
 
     converter.convertWithMetadata.mockRejectedValue(new Error('rate down'));
     const midMonth = new Date(
@@ -531,16 +441,15 @@ describe('CreateAllocatedTransactionBottomSheet — currency create rules', () =
 
     await component['submit']();
 
-    expect(bottomSheetRef.dismiss).not.toHaveBeenCalled();
+    expect(dialogRef.close).not.toHaveBeenCalled();
     expect(component['conversionError']()).toBe(true);
   });
 
   it('should omit metadata fields from payload when inputCurrency equals userCurrency', async () => {
-    const { component, bottomSheetRef, converter } =
-      configureBottomSheetWithCurrency({
-        userCurrency: 'CHF',
-        flagEnabled: true,
-      });
+    const { component, dialogRef, converter } = configureDialogWithCurrency({
+      userCurrency: 'CHF',
+      flagEnabled: true,
+    });
 
     converter.convertWithMetadata.mockResolvedValue({
       convertedAmount: 50,
@@ -566,8 +475,8 @@ describe('CreateAllocatedTransactionBottomSheet — currency create rules', () =
       'CHF',
       'CHF',
     );
-    expect(bottomSheetRef.dismiss).toHaveBeenCalledTimes(1);
-    const dto = bottomSheetRef.dismiss.mock.calls[0][0];
+    expect(dialogRef.close).toHaveBeenCalledTimes(1);
+    const dto = dialogRef.close.mock.calls[0][0];
     expect(dto.amount).toBe(50);
     expect(dto).not.toHaveProperty('originalAmount');
     expect(dto).not.toHaveProperty('originalCurrency');
