+++ frontend/projects/webapp/src/app/feature/budget/budget-details/allocated-transactions/details-dialog/dialog.ts
