+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-details-dialog.service.ts
@@ -20,14 +20,14 @@ import {
   AllocatedTransactionsDialog,
   type AllocatedTransactionsDialogData,
   type AllocatedTransactionsDialogResult,
-} from './allocated-transactions-dialog/allocated-transactions-dialog';
-import { AllocatedTransactionsBottomSheet } from './allocated-transactions-dialog/allocated-transactions-bottom-sheet';
+} from './allocated-transactions/details-dialog/dialog';
+import { AllocatedTransactionsBottomSheet } from './allocated-transactions/details-dialog/bottom-sheet';
 import {
   CreateAllocatedTransactionDialog,
   type CreateAllocatedTransactionDialogData,
-} from './create-allocated-transaction-dialog/create-allocated-transaction-dialog';
-import { computeBudgetPeriodDateConstraints } from './create-allocated-transaction-dialog/budget-period-date-constraints';
-import { CreateAllocatedTransactionBottomSheet } from './create-allocated-transaction-dialog/create-allocated-transaction-bottom-sheet';
+} from './allocated-transactions/create-dialog/dialog';
+import { computeBudgetPeriodDateConstraints } from './allocated-transactions/create-dialog/budget-period-date-constraints';
+import { CreateAllocatedTransactionBottomSheet } from './allocated-transactions/create-dialog/bottom-sheet';
 import {
   ConfirmationDialog,
   type ConfirmationDialogData,
